requirements.txt
absl-py==1.3.0
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
cachetools==5.2.0
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
einops==0.5.0
executing==1.1.1
filelock==3.8.0
flatbuffers==22.9.24
fonttools==4.37.4
google-auth==2.13.0
google-auth-oauthlib==0.4.6
h5py==3.7.0
huggingface-hub==0.10.1
idna==3.4
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.19.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.3
sklearn==0.0
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
timm==0.6.11
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
